# bin/Release/net5.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "RocketElevatorsRESTAPI",
    "version": "v1"
  },
  "paths": {
    "/api/batteries": {
      "get": {
        "tags": [
          "Batteries"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Battery"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Batteries"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Battery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Battery"
            }
          }
        }
      }
    },
    "/api/batteries/{id}": {
      "get": {
        "tags": [
          "Batteries"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Battery"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Batteries"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Battery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Batteries"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/batteries/{requestedInfo}": {
      "get": {
        "tags": [
          "Batteries"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "requestedInfo",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Battery"
              }
            }
          }
        }
      }
    },
    "/api/buildings": {
      "get": {
        "tags": [
          "Buildings"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Building"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Buildings"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Building"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Building"
            }
          }
        }
      }
    },
    "/api/buildings/{id}": {
      "get": {
        "tags": [
          "Buildings"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Building"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Buildings"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/Building"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Buildings"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/buildings/{buildingintervention}": {
      "get": {
        "tags": [
          "Buildings"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "buildingintervention",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Building"
              }
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Battery": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "entity_type": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "date_of_commissioning": {
          "format": "date-time",
          "type": "string"
        },
        "date_of_last_inspection": {
          "format": "date-time",
          "type": "string"
        },
        "certificate_of_operations": {
          "type": "string"
        },
        "information": {
          "type": "string"
        },
        "notes": {
          "type": "string"
        },
        "created_at": {
          "format": "date-time",
          "type": "string"
        },
        "updated_at": {
          "format": "date-time",
          "type": "string"
        },
        "building_id": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
    "Building": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int32",
          "type": "integer"
        },
        "full_name_of_the_building_administrator": {
          "type": "string"
        },
        "email_of_the_administrator_of_the_building": {
          "type": "string"
        },
        "phone_number_of_the_building_administrator": {
          "type": "string"
        },
        "full_name_of_the_technical_contact_for_the_building": {
          "type": "string"
        },
        "technical_contact_email_for_the_building": {
          "type": "string"
        },
        "technical_contact_phone_for_the_building": {
          "type": "string"
        },
        "created_at": {
          "format": "date-time",
          "type": "string"
        },
        "updated_at": {
          "format": "date-time",
          "type": "string"
        },
        "customer_id": {
          "format": "int32",
          "type": "integer"
        },
        "address_id": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
  }
}